Coordinator introduction, reorganizing project.
This commit includes the introduction of a coordinator to manage navigation, also incudes a coordinatorfactory to create the apropiate coordinator.
The focus of this refactory was to create a more robust architecture for the app. In fact includes a reorganization of the project, agrupation in folders represneting functions not techhical features as uncle bob suggest.